misc bug fixes and efficiency gains

+++ Crawler.py
@@ -8,18 +8,18 @@ import datetime
 
 class Crawler(object):
     
-    def __init__(self, dao, expiration_rules, parse_func, extract_content, user_agent='OpenHouseProject.co crawler', sleep_time=.9):
+    def __init__(self, dao, expiration_rules, extract_content, find_pages_to_crawl, user_agent='OpenHouseProject.co crawler', sleep_time=.9):
         self.robots = RobotsCache()
         self.dao = dao
         self.expiration_rules = expiration_rules
-        self.parse_func = parse_func
         self.extract_content = extract_content
+        self.find_pages_to_crawl = find_pages_to_crawl
         self.user_agent = user_agent
         self.sleep_time = sleep_time
         self.parse_errors = {}
         self.work_done = {}
     
-    def crawl_one(self, url):
+    def crawl_one(self, url, verbose=False):
         """ Leverages cache to prevent duplicate crawls
             Checks for a single page in cache, if available
             Crawls for page if not in cache and not expired
@@ -37,7 +37,8 @@ class Crawler(object):
             resp['used_cache'] = False
             allowed = self.robots.allowed(url, self.user_agent)
             if allowed:
-                print 'Crawling', url
+                if verbose:
+                    print 'Crawling', url
                 success = False
                 content = ''
                 while not(success) and failure_tollerance > 0:
@@ -49,7 +50,8 @@ class Crawler(object):
                             content = r.content
                         success = True
                     except requests.exceptions.ConnectionError:
-                        print 'Sleeping due to Connection error'
+                        if verbose:
+                            print 'Sleeping due to Connection error'
                         failure_tollerance -= 1
                         time.sleep(60*10)
                 # TODO: Better error handing for 400, 500, etc.
@@ -72,34 +74,42 @@ class Crawler(object):
         self.parse_errors[err].append(url)
         
 
-    def process_one_url(self, url):
+    def process_one_url(self, url, verbose=False):
         """Get page content and parse
            Handles both content pages (with property details) and directory pages (with links to properties)
            Add other links to queue
            Return parse details
         """
-        summary = {'has_content': False, 'link_count': 0, 'cleared_from_cache': False, 'crawled': False, 'parse': False, 'parse_error': False}
-        crawl = self.crawl_one(url)
+        summary = {'link_count': 0, 'cleared_from_cache': False, 'crawled': False, 'parsed': False, 'parse_error': False}
+        crawl = self.crawl_one(url, verbose)
+        summary['crawled'] = not(crawl['details']['used_cache'])
         content = crawl['content']
-        has_parsed = dao.has_parse(url)
+        if content == None or content == '':
+            summary['parse_error'] = True
+            return summary
+        summary['parsed'] = self.dao.has_parse(url)
         if crawl['details']['cleared_from_cache']:
             summary['cleared_from_cache'] = True
             if has_parsed:
                 self.dao.delete_parse(url)
-                has_parsed = False
+                summary['parsed'] = False
         parse = None
+        saveit = False
         try:
-            if not(has_parsed):
-                print 'Parsing', url
-                parse = self.parse_func(content['content'])
-                summary['parse'] = True
+            if not(summary['parsed']):
+                if verbose:
+                    print 'Parsing', url
+                parse = self.extract_content(content['content'])
+                saveit = True
+                summary['parsed'] = True
         except:
             summary['parse_error'] = True
-            parse = None
-        if parse is not None:
-            print 'saving parse'
+            saveit = False
+        if saveit:
+            if verbose:
+                print 'saving parse'
             self.dao.save_parsed(url, parse)
-        resp = self.extract_content(content['content'])
+        resp = self.find_pages_to_crawl(content['content'])
         links = resp['links']
         nlinks = []
         for link in links:
@@ -129,16 +139,29 @@ class Crawler(object):
         while len(queue) > 0:
             url = queue.pop()
             process = False
-            if not(self.work_done.has_key(url)):
+            haskey = self.work_done.has_key(url)
+            if not(haskey):
                 process = True
-            elif not(self.work_done[url].has_key('used_cache')):
-                process = True
+            elif haskey:
+                hasparsed = self.work_done[url].has_key('parsed')
+                if not(hasparsed):
+                    process = True
+                elif not(self.work_done[url]['parsed']) and not(self.work_done[url]['parse_error']):
+                    process = True
             if process:
                 count += 1
                 if verbose:
-                    print url
-                result = self.process_one_url(url)
-                queue.extend(result['links'])
+                    print('Processing: ' + url)
+                result = self.process_one_url(url, verbose)
+                for link in result['links']:
+                    if not(self.work_done.has_key(link)):
+                        queue.append(link)
+                    else:
+                        wd = self.work_done[link]
+                        if wd.has_key('parsed') and not(wd['parsed']):
+                            queue.append(link)
+                        elif not(wd.has_key('parsed')):
+                            queue.append(link)
                 self.work_done[url] = result['summary']
         end = datetime.datetime.utcnow()
         if verbose:
